This commit address issue #1
Introduced error checking when walking files.

A new WalkerError struct has been introduced and implements the error interface.  The struct stores a slice of errors.
The Walker struct now incorporates the WalkerError struct as an element
Errors are reported via aerror channel and stored on the Walker struct.

Refactored the location of the traversetime bin.  Created an examples folder and placed traversaltime.go in it's own folder within here
Created another example that demonstrates errors.  This is done by attempting to exceed file limits by openng each file as it is traveresed

Updated the README.md with some info on errors